implement shadow tree for file creation
previously, creation of new files was first
attempted in the branch marked for creation (-c
flag) and if that couldn't happen (because of the
path leading to the file not existing in the branch)
the union list was searched in-order until the path
was found, and the file was created there.

this was good enough for the most basic usages, like
the unioning of /bin, in which case if could be
considered useful to route new files into matching
branches; but this behaviour can have unpredictable
results and requires somewhat careful thought to use
advantageously.

forcing the creation of new files in the shadow
tree, i.e.  the one marked with a -c flag, is an
intuitive, easy to understand concept that is also a
very useful default for the kind of usage this
file server was intended for.  going back to our
most basic example of /bin, the user of unionfs will
probably want all the new files and directories created
under /bin to end up in his $home/bin/rc directory,
instead of getting sprinkled out over the system paths.
another very useful usage scenario enabled by this
is build overlays, where the build artifacts can be
transparently collected into a ramfs for improved
speed and ease of cleanup.

this change neccesitates thinking about and eventually
adding a copy-on-write scheme for file modifications,
which would facilitate experimenting with read-only
9front live images with persistable overlays, or
something... :)