Remove flush note retry loop
The retry loop was intended to make sure that we flush the
process, but the approach is very brittle. For one, user none
isn't allowed to send notes at all which would cause this
loop to get stuck forever.

For now, flushing isn't going to work at all if you run unionfs
under 'none'.